requirements.txt
channels==3.0.5
channels-redis==3.4.1
Django==4.1.7
djangorestframework==3.0.14